+++ test/unit/instagram_test.rb
@@ -13,6 +13,16 @@ class InstagramTest < Minitest::Test
     assert_equal "41", browser.version
   end
 
+  test "detects alternate instagram user agent" do
+    browser = Browser.new(Browser["INSTAGRAM_OTHER"])
+
+    assert_equal "Instagram", browser.name
+    assert browser.instagram?
+    assert :instagram, browser.id
+    assert_equal "182257141", browser.full_version
+    assert_equal "182257141", browser.version
+  end
+
   test "detects version by range" do
     browser = Browser.new(Browser["INSTAGRAM"])
     assert browser.instagram?(%w[>=41])