+++ lib/browser/base.rb
@@ -149,6 +149,11 @@ module Browser
       Opera.new(ua).match? && detect_version?(full_version, expected_version)
     end
 
+    # Detect if browser is Sputnik.
+    def sputnik?(expected_version = nil)
+      Sputnik.new(ua) && detect_version?(full_version, expected_version)
+    end
+
     # Detect if browser is Yandex.
     def yandex?(expected_version = nil)
       Yandex.new(ua) && detect_version?(full_version, expected_version)