+++ lib/browser/platform/ios.rb
@@ -4,10 +4,22 @@ module Browser
   class Platform
     class IOS < Base
       MATCHER = /(iPhone|iPad|iPod)/.freeze
-      VERSION_MATCHER = /OS ([\d.]+)/.freeze
+      VERSION_MATCHER = /OS ((?<major>\d)_(?<minor>\d)_?(?<patch>\d)?)/.freeze
 
       def version
-        ua[VERSION_MATCHER, 1] || "0"
+        matches = VERSION_MATCHER.match(ua)
+
+        return "0" unless matches
+
+        versions = [matches[:major]]
+
+        if matches[:patch]
+          versions.push(matches[:minor], matches[:patch])
+        else
+          versions.push(matches[:minor]) unless matches[:minor] == "0"
+        end
+
+        versions.join(".")
       end
 
       def name