+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add more Slack bots.
+- Handle instagram user agents that have a slash instead of a space.
 
 ## 2.6.1