+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add more Slack bots.
 - Handle instagram user agents that have a slash instead of a space.
 - Promote Snapchat to a browser (it was detected as a bot previously).
+- Detect Edge based on Chrome correctly.
 
 ## 2.6.1