+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Handle instagram user agents that have a slash instead of a space.
 - Promote Snapchat to a browser (it was detected as a bot previously).
 - Detect Edge based on Chrome correctly.
+- Improve Yandex detection.
 
 ## 2.6.1